// FE_DEV/src/app/interfaces/consultaAlumnoDTO.ts
export interface consultaAlumnoDTO{
    folio: number;
    nombreAlumno: string;
    apellidoPaterno: string;
    apellidoMaterno: string;
    sexo: string;

    nivel: string;
    grado: string;
    grupo: string;
    turno: string;

    estadoAlumno: string;
    fechaNacimiento: string;
    tipoSangre: string;
    enfermedad: string;

    curp: string;
    email: string;
    becaGobierno: string;
    telCelular: string;

    telefono: string;
    calle: string;
    colonia: string;
    datosAdicionales: string;


}